+++ app/models/user.rb
@@ -1,3 +1,8 @@
 class User < ApplicationRecord
   has_many :events
+  has_and_belongs_to_many :attended_events, class_name: 'Event', join_table: 'event_user'
+
+  def attending?(event)
+    event.attendees.exists?(id)
+  end
 end