+++ config/routes.rb
@@ -6,6 +6,8 @@ Rails.application.routes.draw do
   post '/logout', to: 'sessions#destroy'
   get '/signup', to: 'users#new'
   get '/events/all', to: 'events#all'
+  post '/events/:id/attend', to: 'events#attend', as: 'attend'
+  post '/events/:id/unattend', to: 'events#unattend', as: 'unattend'
 
   resources :users, except: [:index, :new]
   resources :events, except: [:destroy, :edit, :update]