+++ db/schema.rb
@@ -10,7 +10,14 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_07_23_205531) do
+ActiveRecord::Schema.define(version: 2020_07_23_212451) do
+
+  create_table "event_user", force: :cascade do |t|
+    t.integer "event_id", null: false
+    t.integer "user_id", null: false
+    t.index ["event_id"], name: "index_event_user_on_event_id"
+    t.index ["user_id"], name: "index_event_user_on_user_id"
+  end
 
   create_table "events", force: :cascade do |t|
     t.text "description"
@@ -27,5 +34,7 @@ ActiveRecord::Schema.define(version: 2020_07_23_205531) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
+  add_foreign_key "event_user", "events"
+  add_foreign_key "event_user", "users"
   add_foreign_key "events", "users"
 end